rason-forwarder/requirements.txt
loguru==0.7.2
python-decouple==3.8
